requirements.txt
# Flask and Extensions
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0

# Database
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1

# Security
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0

# Data Processing
pandas==2.2.1
openpyxl==3.1.2
XlsxWriter==3.1.9
reportlab==4.1.0

# Utils
Pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
pyngrok==7.1.5
requests==2.31.0
email-validator==2.1.0.post1

# Development
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
python-json-logger==2.0.7 